+++ ElvUI/Modules/Skins/Blizzard/WorldMap.lua
@@ -17,6 +17,7 @@ function S:WorldMapFrame()
 	_G.WorldMapContinentDropDown:Point('TOPLEFT', WorldMapFrame, 'TOPLEFT', 330, -35)
 	_G.WorldMapContinentDropDown:Width(205)
 	_G.WorldMapContinentDropDown:Height(33)
+
 	_G.WorldMapZoneDropDown:Point('LEFT', _G.WorldMapContinentDropDown, 'RIGHT', -20, 0)
 	_G.WorldMapZoneDropDown:Width(205)
 	_G.WorldMapZoneDropDown:Height(33)