+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ### Version 2.07 [ ETA: 🍨 ]
+**• Fixed Soul Bags color issue**  
+**• Blacklisted Honor Event Buff on UnitFrames by default**  
 ___
 ### Version 2.06 [24-05-2021]
 **• Added an option to turn off Quest Icons in Bags**